requirements.txt
antlr4-python3-runtime==4.8
appdirs==1.4.4
audioread==3.0.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorlog==5.0.1
decorator==5.1.1
docker-pycreds==0.4.0
filelock==3.16.1
fsspec==2024.10.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.26.2
hydra-colorlog==1.1.0
hydra-core==1.1.1
hyperlink==17.3.1
HyperPyYAML==1.0.0
idna==3.7
importlib_metadata==8.5.0
importlib_resources==6.4.0
iopath==0.1.10
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.41.1
memory-profiler==0.61.0
msgpack==1.1.0
numba==0.58.1
numpy==1.24.4
omegaconf==2.1.2
opencv-python==4.9.0.80
packaging==24.1
pesq==0.0.4
pillow==10.3.0
platformdirs==4.3.6
pooch==1.8.2
portalocker==2.10.1
protobuf==4.25.3
psutil==5.9.8
pycparser==2.22
PyYAML==6.0.1
requests==2.31.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.10.1
sentry-sdk==1.45.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
soundfile==0.12.1
sox==1.5.0
soxr==0.3.7
tabulate==0.9.0
termcolor==2.4.0
threadpoolctl==3.5.0
timm==0.9.5
torch @ file:///home/hhdx/PycharmProjects/torch%201.12.1/torch-1.12.1%2Bcu113-cp38-cp38-linux_x86_64.whl#sha256=4adf483ac2d047534a7d023f0022bd8694d87627068ad6dddf186cb3273bbfa2
torchaudio @ file:///home/hhdx/PycharmProjects/torch%201.12.1/torchaudio-0.12.1%2Bcu113-cp38-cp38-linux_x86_64.whl#sha256=07324a67c069a986bb991c509c9442e8fce72326941e55fa4738d88cbb806047
torchvision @ file:///home/hhdx/PycharmProjects/torch%201.12.1/torchvision-0.13.1%2Bcu113-cp38-cp38-linux_x86_64.whl#sha256=899cac298d2a7cf6a8ca62d3ede2a7d3f50b86027f8be2d15639baf74da5d8f0
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
wandb==0.16.6
yacs==0.1.8
zipp==3.20.2
